requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
arrow==1.2.3
    # via django-q
asgiref==3.5.2
    # via django
blessed==1.19.1
    # via django-q
boto3==1.24.81
    # via -r requirements.in
botocore==1.27.81
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via whitenoise
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.1.1
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-import-export
    #   django-model-utils
    #   django-picklefield
    #   django-q
    #   django-sesame
    #   django-storages
django-constance[database]==2.9.1
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-import-export==2.9.0
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-picklefield==3.1
    # via
    #   django-constance
    #   django-q
django-q==1.3.9
    # via
    #   -r requirements.in
    #   django-q-email
django-q-email==5.0.1
    # via -r requirements.in
django-sesame[ua]==3.1
    # via -r requirements.in
django-storages==1.13.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.1.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markuppy==1.14
    # via tablib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
psycopg2-binary==2.9.3
    # via -r requirements.in
pymemcache==3.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
pyyaml==6.0
    # via tablib
redis==3.5.3
    # via django-q
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   blessed
    #   pymemcache
    #   python-dateutil
sqlparse==0.4.3
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
ua-parser==0.16.1
    # via django-sesame
urllib3==1.26.12
    # via botocore
wcwidth==0.2.5
    # via blessed
whitenoise[brotli]==6.2.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
